Refactor code into Ui, Storage, Parser and TaskList classes
